+++ app/models/design_option.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+# Defines the position of sections and questions on designs. Design options also
+# have associated branching logic, and can be set as recommended or required
 class DesignOption < ActiveRecord::Base
   # Model Validation
   validates :design_id, presence: true