+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added coffeelint configuration file for CodeClimate
 - Continued cleanup and refactoring, and additional testing of controllers
 - Refactored design imports and reports to dedicated controllers
+- Removed design JSON exports and imports
 
 ## 0.35.1 (March 1, 2016)